fix: moved unit tests inside the rust implementation file
